Remove function to work with files
