+++ lib/ansible/config/manager.py
@@ -29,7 +29,7 @@ from ansible.module_utils._text import to_text, to_bytes, to_native
 from ansible.module_utils.parsing.convert_bool import boolean
 from ansible.parsing.quoting import unquote
 from ansible.utils import py3compat
-from ansible.utils.path import cleanup_tmp_dir, makedirs_safe, unfrackpath
+from ansible.utils.path import cleanup_tmp_file, makedirs_safe, unfrackpath
 
 
 Plugin = namedtuple('Plugin', 'name type')
@@ -110,7 +110,7 @@ def ensure_type(value, value_type, origin=None):
                 makedirs_safe(value, 0o700)
             prefix = 'ansible-local-%s' % os.getpid()
             value = tempfile.mkdtemp(prefix=prefix, dir=value)
-            atexit.register(cleanup_tmp_dir, value)
+            atexit.register(cleanup_tmp_file, value, warn=True)
 
         elif value_type == 'pathspec':
             if isinstance(value, string_types):
